Move HistoryStoreSpy to its own file
