+++ web-browser/web-browserTests/Window/WindowViewAdapterTests.swift
@@ -87,38 +87,6 @@ class WindowViewAdapterTests: XCTestCase {
     }
 }
 
-private class HistoryStoreSpy: HistoryAPI {
-    enum Message: Equatable {
-        case save(_ url: URL)
-        case getPages
-        case deletePages
-        case deleteAllPages
-    }
-
-    var receivedMessages = [Message]()
-
-    func save(_ page: WebPage) {
-        receivedMessages.append(.save(page.url))
-    }
-
-    func getPages() -> [WebPage] {
-        receivedMessages.append(.getPages)
-        return []
-    }
-
-    func getPages(by searchTerm: String) -> [WebPage] {
-        return []
-    }
-
-    func deletePages(withIDs ids: [UUID]) {
-        receivedMessages.append(.deletePages)
-    }
-
-    func deleteAllPages() {
-        receivedMessages.append(.deleteAllPages)
-    }
-}
-
 class SafelistStoreSpy: SafelistAPI {
     enum Message: Equatable {
         case isRegisteredDomain(_ domain: String)
